Move `savePost` into the library
