+++ CHANGELOG.md
@@ -1,0 +1,10 @@
+## [Unreleased]
+
+### Added
+- New function, `IO.savePost`, that writes a given post onto the disk, complete
+    with a "front matter" understood by Hakyll
+
+### Changed
+### Deprecated
+### Removed
+### Fixed
